refactor(contentful): migrate to schema customization API
